engine/pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["Maxence Frenette <maxence.frenette@gmail.com>"]
license = "GPL-3.0-or-later"
packages = [
    {include = "engine", from = "."}
]

[tool.poetry.dependencies]
python = "^3.12"
python-chess = "^1.999"
training = {path = "../training", develop = true}
torch = "^2.6.0"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
